Multi-regions Support
Add support for basic multiple regions, that is to say, many OpenStack
with a shared Keystone (same users). The shared Keystone is deployed
into one region, for instance RegionOne. Services of other regions
(e.g., RegionTwo, ..., RegionN) have an access to this Keystone. This
support assumes that the operator knows the name of all OpenStack
regions in advance, and considers as many Kolla runs as there are
regions.

The new variable, multiple_regions_names, contains the name of
regions. It is only needed by the region that includes Keystone and
creates keystone endpoints so that services of other regions can
connect to Keystone. The multi-regions.rst explains how to perform a
multiple regions deployment.